feat: add watch progress endpoint and update duration handling in movie entities

+++ src/movie-service/src/main/java/com/pbl6/cinemate/movie/service/impl/MovieTranscodeServiceImpl.java
@@ -43,7 +43,7 @@ public class MovieTranscodeServiceImpl implements MovieTranscodeService {
             return;
         long dur = metadata.durationSeconds();
         if (dur > 0 && dur <= Integer.MAX_VALUE) {
-            movie.setDuration((int) dur);
+            movie.setDuration(dur);
             repo.save(movie);
         }
     }
